requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
Markdown==3.3.6
psycopg2-binary==2.9.3
Pygments==2.12.0
pytz==2022.1
sqlparse==0.4.2
